Do not change repository branch if commit is constant (#11)
If a commit is not changing, don't update the underlying repository
branch. This allows a user to checkout a HEAD or different branch than
the root branch, without a `gordion -u` moving it back as long as the
commit is not changing. It also saves us from needing to call fetch in
the semi-nominal workflow where a child repository does not have the
same branch as the root repository. In this case, the previous behavior
was to fetch the remote to see if it contains the root branch every time
we update, which takes time and is annoying for a common behavior. Now,
it can be expected that if a user calls `gordion -u` twice in a row, the
second time will do nothing and be fast.